USe cleanup logic also for BPEL System Tests.
